+++ packages/esense_flutter/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.6.1
+* fixed bug in sampling rate configuration on Android
+* small improvement to API doc and example app
+
+## 0.6.0
+* moved permission handling to app level
+* improved on iOS implementation
+* improvement to exampel app
+* improvement to README
+
+## 0.5.0
+* upgrade to Android embedding v2
+
 ## 0.4.3
 * fix of error in setSamplingRate method on iOS